+++ Assets/Prefabs/Game System.prefab
@@ -45,12 +45,24 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 5fcea4a4b04c3574cb2edbde059de83f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  reputation: {fileID: 11400000, guid: 8ea7d138f1c905945908b28ebbbb7648, type: 2}
   playerPrefab: {fileID: 2462193850337484257, guid: 270f7328ee227ae44aaac2aa921cf183, type: 3}
-  laserPrefab: {fileID: 2254531325587726217, guid: 15d30101cc2178d488611d4a1cb65218, type: 3}
   playerShip: {fileID: 0}
+  playerLaserPrefab: {fileID: 0}
   playerSpawnPosition: {x: 0, y: 0, z: 0}
   playerSpawnRotation: {x: 0, y: 0, z: 0, w: 1}
   isNewGame: 1
+  hasPlayedTutorial: 0
+  onMenu: 0
+  tutorialScenario: {fileID: 11400000, guid: 14d8742131d858c4691c5442129473c0, type: 2}
+  quickPlayScenario: {fileID: 11400000, guid: e371798ec0652a54f8c19df950343949, type: 2}
+  scenarios:
+  - {fileID: 11400000, guid: 96251ac8dc2c5c24e94c35d1e2b7d06d, type: 2}
+  - {fileID: 11400000, guid: 10f49a4a1bc257f4893cbba1a3563944, type: 2}
+  - {fileID: 11400000, guid: 74a5ad1a41c9fca46a4cd0c7f92bbbc0, type: 2}
+  - {fileID: 11400000, guid: 25107e128c8d4b44aae9e282179587ba, type: 2}
+  currentScenario: {fileID: 0}
+  portalPosition: {x: 25, y: 25, z: 25}
 --- !u!114 &2855395254434112166
 MonoBehaviour:
   m_ObjectHideFlags: 0
