+++ Assets/Resources/Scriptable/Player Reputation.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8ea7d138f1c905945908b28ebbbb7648
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
