+++ Assets/Scripts/Model/AI/AIState.cs
@@ -9,7 +9,7 @@ namespace Model.AI
 		Roam,    // The AI is wandering around
 		Seek,    // The AI is chasing or attacking the player
 		Flee,     // The AI is running away (e.g., when health is low)
-		AllyAssist // The AI is assisting the player
-
+		AllyAssist, // The AI is assisting the player
+		Attack
 	}
 }
