+++ Assets/Scripts/Model/AIBehavior.cs
@@ -7,14 +7,17 @@ namespace Model
     [CreateAssetMenu(fileName = "New AI Behavior", menuName = "AI/Behavior")]
     public class AIBehavior : ScriptableObject
     {
+        public enum Faction { Colonial, Earth, Pirates, Solo }
+        public enum GroupMode { None, Formation }
 
         [Header("General Settings")]
+        public Faction faction;
         public float aggressionLevel; // How aggressive this faction is
         public float reputationThreshold; // Reputation level required to avoid attacks
+        public float allyReputationThreshold = 70f; // Reputation level required to ally with the player
         public float allyAssistRange = 20f; // Range within which the AI will assist the player
         [Range(0, 2)] public float missionResponseAggression = 1f;
         public float detectionRadius = 25f;
-
 
         [Header("Movement")]
         public float roamSpeed;
@@ -24,6 +27,15 @@ namespace Model
         [Header("Combat")]
         public float attackRange;
         public float attackCooldown;
+
+        [Header("Group Behavior")]
+        public GroupMode groupMode = GroupMode.None;
+
+        [Header("Aggression Tuning")]
+        public float aggressionCalmValue = 1f;
+        public float aggressionAverageValue = 4f;
+        public float aggressionAggressiveValue = 10f;
+
     }
 
 }
